[go] Meta Quest support - [1/n] - Add product flavors (#37743)
# Why

This is the first part of PRs that add support for the Meta Quest to
Expo Go. We need slightly different android manifests for mobile devices
and the quest. The best way of getting that is adding product flavors.

# How

Added mobile and quest flavors. 

Updated `eas.json` of expo go to run gradle commands with flavors

The common manifest in `template-files/android/AndroidManifest.xml` now
contains only properties shared across the platforms. The manifest in
`template-files/android/mobile/AndroidManifest.xml` contains properties
used on mobile but not used on the quest. The same for
`template-files/android/quest/AndroidManifest.xml`

Modified the generate-dynamic-macros script to create the desitination
directories if they don't exist (`/quest` and `/mobile` won't exist).
Also updated the `WRITE_CONTACTS` permission injection to also check for
the `mobileDebug` variant.

# Test Plan

Tested on a local build of Expo-Go with the mobile and quest variants.