[core] Meta Quest support - [2/n] Add VR utilities (#37744)
# Why

We need a place to store horizon os permissions that we are requesting
as well as a common way to check if the current device is a quest
headset.

# How

Added VRUtilities.kt file that stores the camera permission constant as
well as a simple function that allows us to check if the current device
is a quest at runtime.

# Test Plan

Tested in Expo Go when adding support for the Quest