[go] Add Brewfile + instructions to Expo Go readme
Why
===
Expo Go builds Hermes from source, which requires cmake and ninja (and git, but that's obviously installed in order to clone this repo - so we don't want to touch that).

Other projects under `apps` do not link to react-native-lab the same way Expo Go does so I updated only the Expo Go instructions.

How
===
Updated the Expo Go readme to tell people to install tools using Homebrew with `brew bundle`. This uses the dependencies specified in a Brewfile introduced by this commit, which documents why we're adding cmake and ninja in a way that is aligned with the Hermes build instructions.

Test Plan
===
Ran `brew bundle` in the root directory. I didn't try building Expo Go but Hirbod mentioned that installing cmake 4.0.3 (latest in the brew tap) got Expo Go building for him successfully.