codereview: refactor apps argument
The apps argument passed to a tool might be a bit big to pass as a stringified json on the command line.

This commit fixes this by changing the apps argument to be the metro server origin so that the extension can use the `/json/list` endpoint by itself to get a list of connected apps.

The metro server origin should be available for both cli extensions and also future mcp extensions.