[sdk54][0.81.0.rc-3] cleaned up iOS header files (#38441)
# Why

Quite a few of our iOS header files were importing umbrella files directly. This is not a public and available API and will cause issues when we start precompiling e-m-c and other expo modules.

# How

Removed includes using umbrella headers and swapped with correct includes or forward declared classes where applicable. There are some places where it is not doable, but this will be fixed later when continuing to work on precompiled EMC.

# Test Plan

Tested in BareExpo on iOS running both with/without precompiled React Native binaries.

# Checklist

- [x] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).