[android][go] Add FAB for old dev-menu (#38371)
# Why

I've added FAB support to the new dev-menu, but we won't use it in the
SDK 54 version of expo go. Therefore I've added it to the current
expo-go dev-menu.

This PR depends on the changes introduced in there:
https://github.com/expo/react-native/pull/41
Once above is merged we should change the commit hash for the
react-native-labs submodule in this PR.

# How

Copied the FAB implementation to expo-go, updated it slightly to work
with the setup in Expo Go. The FAB is added above the app content, but
below the dev menu, so it stays in the background once the dev menu is
pulled up.
I have also added a button to show/hide the button into the dev-menu.

# Test Plan

Tested in Expo Go running NCL on an Android 16 device.