fix(expo,babel-preset-expo): Fix/Remove new peer dependencies on `expo` (and bump `@babel/core` in monorepo) (#38562)
# Why

Fixes added peer dependencies by: #36405

Generally, `expo` is installed directly, so the only way peer
dependencies by it are fulfilled is if a user directly installs
dependencies.
It's not desirable and not the case that `react-refresh` and
`metro-runtime` are installed in our templates and generally we don't
want to add them as they're not the user's concern.

# How

To fulfill the change of `expo/src/async-require` depending on
`react-refresh`, we move it to be a dependency in `expo`. The dependency
in `babel-preset-expo` is turned into a peer dependency. While both can
be a dependency, we want the runtime versions to match up, and if we
don't expose the runtime from `babel-preset-expo`, a peer here should
suffice.

`metro-runtime` in `expo/src/async-require/hmr.ts` has been replaced
with `@expo/metro/metro-runtime`. This also isn't a desirable peer, as
it calls for the user to install it, potentially installing a misaligned
version. For those reasons, it's been fully removed.

This PR also adds two missing dev dependencies on `@babel/core` (~~which
also happened to bump some versions, which is technically a drive-by
update, but I don't see a reason to prevent those updates manually~~;
Edit: Reverted the soft-bump of deps to prevent excessive drive-by test
updates being needed). This should mute more in-monorepo warnings of
missing peer dependencies in `expo/expo` that only apply to our
monorepo. Fewer warnings would've given us a higher chance of catching
the two above issues earlier.

# Test Plan

- n/a: Resolves install issues only reported in warnings

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)