fix(babel-preset-expo): Add explicit optional reverse peer dependency on `expo` for `expo/config` loading (#38600)
# Why

A common failure case with Yarn Berry is that dependencies are hoisted
past their dependents. This may happen more frequently with it because
it opts to have a bottom-up hoister that may abort hoisting if any
check-condition fails. Specifically, it may decide to start hoisting
`babel-preset-expo` past `expo` (in a monorepo), but not hoist `expo`.

The dependency on `expo/config` may also be an issue in
`babel-preset-expo`, if `babel-preset-expo` is used independently. We
can gracefully skip this plugin as well for those cases. This is
technically not supported, but the plugin feels like it shouldn't be in
`babel-preset-expo` in that case and instead in `@expo/metro-config` or
`@expo/cli`.

Either way, we'd like to establish a correct (reverse) dependency chain
from `babel-preset-expo` to `expo`. In this case it's marked as optional
since this is a bit of an edge case.

# How

- Add optional (reverse) peer dependency on `expo` to
`babel-preset-expo`
- Refactor `expo-inline-manifest-plugin` to lazily require `expo/config`
and to handle errors by defaulting to skipping its logic

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)