[expo-blob] Added ExpoBlob package (#38366)
# Why
React-native Blob implementation is not compliant with the
[w3c](https://w3c.github.io/FileAPI/#dfn-Blob) Blob interface and is not
efficient.
We wanted to implement the full interface so that it works like the web
implementation, but also keep it simple.

# How
We added an expo-blob package which uses Kotlin and Swift in native
implementation consistent with
[w3c](https://w3c.github.io/FileAPI/#dfn-Blob) documentation.

# Test Plan
We have written several own tests and copied most of the tests from
[wpt](https://github.com/web-platform-tests/wpt/tree/master/FileAPI/blob)
and made sure that both the Kotlin and Swift implementations pass these
tests. The tests from wpt which we are not using:
- DOM related tests - not applicable for react native,
- Float16Array test -  no support in react native.

# Checklist
- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [x] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)

---------

Co-authored-by: Jakub Grzywacz <jakub.grzywacz@swmansion.com>
Co-authored-by: Jakub Grzywacz <kontakt@jakubgrzywacz.pl>
Co-authored-by: HubertBer <115428831+HubertBer@users.noreply.github.com>
Co-authored-by: HubertBer <bernackihubert03@gmail.com>