[docs] Update autolinking docs and references (SDK 54) (#38691)
This applies and documents changes for #38635 and #38282.

The main page that has changed is the "Expo Autolinking" page. I've
added a new section that explains how modules are resolved in an app and
added more references to the React Native module linking behavior.

The React Native linking behavior is subject to some newer changes (and
a new config flag), so we can't leave it out anymore. It'll also be
supported by every SDK in our support window (52–54).

I've also added notes where SDK 54 has changed the default behaviour
(two places) in the Configuration reference section.

Minor fixes in this PR include:
- Removing `expo-application` from list of expo dependencies mentioned
in a doc
- When we're telling users how to set up a monorepo for Expo modules,
tell them to add a regular dependency rather than altering the
`searchPaths` (this worked prior to SDK 54 too)
- Add a missing cross-link for `buildFromSource`

---------

Co-authored-by: Aman Mittal <amandeepmittal@live.com>