[expo-router] fix preloading of protected routes (#38789)
# Why

When `<Link prefetch>` or `Link.Preview` were used to the protected
route, then an error was thrown. This was caused by the mismatch of our
linking configuration and react-navigation's state - the protected
routes are not present in react-navigation, but are still present in
linking config. The mismatch was coming from `getStateFromPath` function
in `linkTo` function. This function was creating a state, which was not
valid in react-navigation.

This is just a fix to the issue, but maybe we should rethink how we
define protected routes, and wether or not they should be present in
linking config.

# How

Disable preloading of routes that are not available in react-navigation
navigator

# Test Plan

1. Unit tests - all current tests should pass
2. Manual testing

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)