[ios][expo-router] Suppress native link preview for Apple TV (#38814)
# Why

The iOS native link previews in expo-router require webviews, so they don't compile for tvOS. We need to suppress them in `native.ts`, so that we don't get unnecessary warnings in TV builds.

# How

Check that `Platform.isTV` is not true for native links. This will only be set in builds using `react-native-tvos` and running on Apple TV or Android TV.

# Test Plan

Tested locally to verify that the warnings do not occur after this change. There should be no effect on non-TV platforms. CI should pass.

# Checklist

- [x] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [x] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)