[tools][publish-packages] Fix forced template release version (#38805)
# Why

When publishing `expo` package, `expo-template-bare-minimum` update is
forced, regardless of selected options to keep the integrity, however
it's version is based on `expo` update version. When the version is not
a major, template version should be patch.

# How

Updated the condition for release type. 

# Test Plan

Publish `expo` package as prerelease or minor without any additional
packages and check the logs.
The output should be
```
📦 expo-template-bare-minimum is required to be published with expo package, will be published as 54.0.1.
```
instead of
```
📦 expo-template-bare-minimum is required to be published with expo package, will be published as 55.0.0.
```