[docs] Update upgrade steps for reliability (#38827)
# Why

In the docs, since we have the ability to show tabs, there's no need to
recommend users to use `npx expo install expo`. I think it's a little
redundant, and given we instruct users to run `expo install --fix`
(which they have to after) this doesn't add any clarity. In fact, it can
fail in some cases.

We were also missing a step for users to run `expo-doctor` after
uprgading. This should, imho, always be done, as there's new problems it
may catch after an SDK upgrade, since it does dependency checks.

Might be controversial, and we haven't been super consistent in giving
people these instructions, but this is the most reliable sequence for
upgrading, as far as I'm aware. The only gotcha here would be that in a
Beta, `npx expo-doctor` would have to be replaced with `npx
expo-doctor@next`

# How

- Fan out install instructions for SDK versions
- Remove `@latest` variant, since I don't think we should recommend a
blind upgrade to latest
- Add `expo-doctor` after `npx expo install --fix`

---------

Co-authored-by: Aman Mittal <amandeepmittal@live.com>
Co-authored-by: Kadi Kraman <kadi@expo.io>