Remove onlyProjectDeps
This is now not an option any longer. This only made sense when the
searchPaths had defaults that included the entire monorepo. But this
behaviour is now basically the default with new resolution and the
option is ineffective. It was also unclear how it'd be set given that
our build calls don't have a corresponding option and it wasn't
documented.