[expo][android][1/n] Use React Native's edge-to-edge - Respect `enableEdgeToEdge` in Expo projects (#38734)
# Why

Currently Expo projects don't respect the new `enableEdgeToEdge` Gradle
property. That's because we override the
`ReactActivityDelegate.onCreate` and don't call the original method,
where the property is read and applied.

# How

Got React Native's internal edge-to-edge methods with reflection and
launched them in our `onCreate` implementation.

# Test Plan

Tested in BareExpo in Android SDK 34 emulator.