[expo-status-bar][android][3/n] Use React Native's edge-to-edge - Remove `react-native-edge-to-edge` (#38768)
# Why

We don't want to have `react-native-edge-to-edge` as a transitive
dependency. It will cause issues by always enabling edge-to-edge

# How

Rolled back to our previous implementation instead of using `SystemBars`
from `react-native-edge-to-edge` turns out, our old methods still work
with edge-to-edge enabled. I've left warnings for properties that do not
work with edge to edge enabled (they also didn't work for SystemBars)`

# Test Plan

Tested in BareExpo on Android