[expo][android][2/n] Use React Native's edge-to-edge - Drop `react-native-edge-to-edge` dependency (#38767)
# Why

We are moving onto using the new `enableEdgeToEdge` property in react
native. This means that we can drop the dependency on
`react-native-edge-to-edge`.

# How

- Removed the `react-native-edge-to-edge` dependency from `expo`
- Our default edge-to-edge plugin now only sets the gradle property to
enable/disable edge-to-edge
- Added `enforceContrast` property to `androidNavigationBar` to fill a
functionality void left after removing react-native-edge-to-edge
- Updated the tests - Removed some no longer necessary tests and added
new ones for the `enforceContrast` property
- Not setting `edgeToEdgeEnabled` will now enable edge-to-edge in users'
projects instead of disabling it
- Renamed the plugin directory from `react-native-edge-to-edge` to
`edge-to-edge`.

# Test Plan

Tested in a small fresh project in the monorepo and in BareExpo