[expo-navigation-bar][android][4/n] Use React Native's edge-to-edge - Remove react-native-edge-to-edge #38768 (#38769)
# Why

Same as for `expo-status-bar` we want to remove the
`react-native-edge-to-edge` dependency to avoid issues.

# How

Removed the `SystemBars` our previous implementations seem to work fine
for everything that the ones from `SystemBars` worked for. Kept the
`react-native-is-edge-to-edge` dependency to be able to check whether
edge-to-edge is enabled.

# Test Plan

Tested in BareExpo on Android