fix(expo-router): Retype internal `ViewSlot` and remove `SafeAreaViewSlot` (RN warning) (#38899)
# Why

This was still relying on the `SafeAreaView` import. This wasn't
actually used by value, but still triggered the deprecation warning.
I'm not sure what the `ref` type change in `SafeAreaViewSlot` was used
for but we can't actually guarantee a specific value there unless we
know the child. If we know the child, we're not using the `ref`
(anymore?), so we don't need it, and also shouldn't make cases like
these reusable.

# How

`SafeAreaViewSlot` is marked as deprecated, so we don't use it anymore
internally. It looked like we may kept it around for a reason, so I'm
still keeping it around for now.

To account for this, I changed the `Slot` types entirely to a generic
`interface Slot<Props = ViewProps, Ref = Component<ViewProps>> {}`. The
`Ref` generic type still isn't accurate, but this type allows users to
override this to be more accurate.

# Test Plan

- Manually test `Tabs` layout and a common layout that uses `Slot` by
default

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)