[docs] Replace "reflect [...] automatically" in getting started tutorial (#38956)
# Why

Feedback from Discord:
https://discord.com/channels/695411232856997968/1406292483469742183/1406292483469742183

It's inaccurate to say "reflect [...] automatically". Because this is a
tutorial for beginners it's helpful to be more precise, and we're trying
to explain the conditions for fast refresh here.

# How

- Replace "automatically" with "Once you save your changes,"
- Replace "reflect" with "sent and applied to the running apps connected
to the development server"

This way this represents the conditions under which fast refresh works
more accurately.