Apply module remapping in supervising-transform-worker
The supervising transform worker wraps around the user's custom
transformer. When this custom transformer requests modules that match
a limited set of SDK packages that are fixed to the `expo` package (i.e.
the SDK version) we redirect the request to fixed paths.

This ensures that a transform worker doesn't break even if it imports
from "illegal" paths.

This applies to:
- `@expo/metro-config` (self)
- `@expo/metro`
- `metro-transform-worker`
- `metro`

These packages should never mismatch with the SDK's version of them.