[video] Add `keepScreenOnWhilePlaying` property (#37137)
# Why

Currently on iOS the player will keep the screen on while a video is
playing, on Android playing a video has no effect on screen going to
sleep.
This PR unifies the behavior and adds a new property:
`keepScreenOnWhilePlaying` to allow users to control this feature.

# How

Added `keepScreenOnWhilePlaying`, enabled it by default on Android. Also
the fullscreen player on Android will always stay on.

# Test Plan

Tested in BareExpo on physical Android and iOS devices.