[doctor] Inform users about corrupted `node_modules` folders via autolinking duplicates check (#39026)
# Why

Context from Discord:
https://discord.com/channels/695411232856997968/1407839724193972244

A user reported that Bun failed to clean up after itself. It kept around
old duplicate dependencies (presumably from when prior conflicts existed
before `expo install --fix`). When it fails to clean up after itself,
duplicates exist in the `node_modules` but not in the lockfile.

This is recognisable by seeing duplicates with identical versions.
Packages with identical versions should always be deduplicated.

# How

Check if a resolution has duplicates that are all the same version as
the base resolution. If so, the package manager failed to hoist (or
rather, most of the time, failed to delete old module folders)

We should tell the user to delete their `node_modules` folder(s) and
reinstall in that case.