[video][ios] Fix `sourceLoad` event not being emitted (#39023)
# Why

The `sourceLoad` event was not being emitted. This is because we were
checking the status of the player, which was still loading, instead of
checking the status of the PlayerItem, which was already loaded.

# How

We check the status of the playerItem instead of the player.

# Test Plan

Tested in BareExpo on an iPhone 13