[video][android] Always enable native controls in fullscreen (#39015)
# Why

It is possible to enter fulllscreen with native controls disabled on
Android. This is not in line with the iOS behaviour, where the native
controls are always enabled. It is better to keep them always enabled,
it's not possible to add custom controls in fullscreen and displaying
just video may be confusing for end users.

Fixes https://github.com/expo/expo/issues/38968

# How

The native controls in the fullscreen player activity view are now
always visible

# Test Plan

Tested in BareExpo on Pixel 8 running Android 16