[cli] Remove force-flipping of `EXPO_USE_FAST_RESOLVER` flag for experiments (#39073)
The fast resolver is different from Metros' resolver and predates
`package.json:exports` support. As such, it isn't behaving the exact
same in all cases. This will be difficult to align and maintain and it's
hard to tell, if we'll have the capacity and coverage to maintain two
versions, rather than improving/replacing one and only have one at a
time.

It's not necessary for the other experimental features any longer, so we
can disable it across the board, and only switch it on if its
environment variable is flipped manually.

Docs have been updated to stop indicating it'll become the default.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>