feat(prebuild): Set splash background color to default system background (#39047)
# Why

- Apple HIG specifies that splash screens should be as invisible as
possible.
https://developer.apple.com/design/human-interface-guidelines/launching
- This PR sets the default view background color to the Apple system
background color so it adapts to different platforms by default. This
can be overwritten with the splash plugin, but if no plugin is provided
then you generally get a white background in light mode and a black one
in dark mode (depending on if you allow dark mode).

# How

<!--
How did you build this feature or fix this bug and why?
-->

# Test Plan

- `yarn prebuild` in router-e2e -> open in Xcode