feat(expo): add structuredClone to native apps (#37503)
# Why

`structuredClone` is a widely supported API in the JavaScript world,
added in 2022. It's part of the HTML spec which means it's in a bit of a
weird place for React Native since we don't support HTML. However
runtimes like Node.js, Bun, and Deno all support it to.
`structuredClone` is becoming more popular with the rise of AI apps, and
the missing API leads to issues like this
https://github.com/vercel/ai/issues/4696
This PR proposes we add the API via the package @ungap/structured-clone
and make it globally available for all native apps.

We should consider splitting the polyfills out in the near future and
making them standalone built-ins to reduce bundling time and complexity.
Right now we'd need to be able to roll up all of expo → react-native in
order to split these builtins out of the standard bundle pipeline.

# Test Plan

- Added a test like the TextEncoder test to ensure the polyfill is
installed.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>