[video][android] Bump media3 to 1.8.0 (#39184)
# Why

Fixes https://github.com/expo/expo/issues/39086

We are using an outdated media3 version from 10 months ago. We can bump
it from 1.4.0 to 1.8.0.
We are using a modified version of `DefaultLoadController.java` from
media3 source, converted to `VideoLoadController.kt` to provide
buffering options to our users. While writing this PR I've found out
that the upgrade process with the current arrangement is a bit of a
nightmare when changes to `DefaultLoadController` are introduced in
media3. That is, because it hard to see what actually changed and needs
to be updated between the versions. I've decided to change the approach
- see below:

# How

- Bumped media3 to 1.8.0 version. 
- DefaultLoadController is now copied and left in java, the few simple,
minimal changes (making stuff non-private) are explained in the note at
the top
- `VideoLoadController` inherits from `DefaultLoadController` providing
support for `BufferOptions`
This should make the upgrade process in the future easy.

# Test Plan

Tested in BareExpo on Android