[ci] fail when attempting to run nonexistent test suite (#39122)
# Why

Some of the test suite names have no tests actually associated with them, and so the test suite does actually run any tests. Because no tests fail, previously this would be reported as success. This PR makes it so that this case is an error.

# How

- Remove the test names for which there are no tests (Permissions, HTML...)
- in `apps/test-suite/screens/TestScreen.js`, report an error when there are no matching tests

# Test Plan

- test suite passes in CI

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [ ] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)