[android] [expo-clipboard] Obfuscate filenames in clipboard cache (#39124)
# Why
Expo-clipboard uses a custom content provider to share images via the
clipboard. This provider is exported, so any app can query for it's
content. Since the images are always stored as `copied_image.<ext>` it's
possible to bypass the clipboard by querying for the image directly.
This PR randomizes the names of stored images to prevent this.

# How
This PR changes how the image file is created and replaced:
- We use a hash (sha-256) as the filename for the image we store in the
cache.
- When a new image is added, we clear the entire clipboard cache
directory. This is necessary since the images no longer have the same
name and aren't replaced when a new one is added. This keeps the cache
ephemeral and prevents bloat.

# Test Plan
1. Ran Expo Test Suite.
2. Used the api to get and set an image.
3. Logged the clipboard content, stored image path and cache directory
contents to check if it is empty before a new image is added.
4. Validated that the image was accessible before the changes in this
PR.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [x] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)

---------

Co-authored-by: Wojciech Dróżdż <31368152+behenate@users.noreply.github.com>