[cli] Fix unnecessary version warning when prebuilding prerelease versions (#39298)
# Why

ENG-17186

We show an unnecessary warning when the prerelease version of expo is
explicitly specified:
in package.json e.g. :
```
"expo": "54.0.0-preview.10",
```
when the recommended version is `~54.0.0-preview.10`, will result in the
following warning:
`› Using expo@54.0.0-preview.10 instead of recommended
expo@~54.0.0-preview.10.`


# How

After some research, turns out that semver has a `includePrerelease`
that makes it work the same as the npm versioning.
See more
[here](https://github.com/Masterminds/semver?tab=readme-ov-file#working-with-prerelease-versions)

# Test Plan

Ran `npx expo prebuild` in a project.