[expo][iOS] add isLiquidGlassAvailable function (#39319)
# Why

There’s an alternative version of this PR that adds the check to
expo-application. I’m not entirely sure which package is best suited for
this check. https://github.com/expo/expo/pull/39289


https://linear.app/expo/issue/ENG-17200/add-hook-to-detect-if-liquid-glass-is-enableddisabled

Detection of Liquid Glass can be done at the start of the application.
There is no way to disable new Liquid Glass components (e.g., TabBar) at
runtime. The only thing users can change is whether the Liquid Glass
effect is visible in the new components.

For example, when the Reduced Transparency accessibility setting is
enabled, the new components will render with a plain background. This
check can should be performed via the AccessibilityInfo API, as it is
clearly an accessibility-related feature.

# How

1. Check Xcode version – Ensure the app is compiled using Xcode 26
(compiler ≥ 6.2). If the app is compiled with an older Xcode, Liquid
Glass will not be available.
2. Check OS version – Ensure the app runs on iOS/macOS/tvOS 26 or later.
Liquid Glass is only available on iOS 26+.
3. Check design compatibility – If the
[UIDesignRequiresCompatibility](https://developer.apple.com/documentation/BundleResources/Information-Property-List/UIDesignRequiresCompatibility)
key is set to `true`, Liquid Glass will be disabled.

# Test Plan

1. Test if the new constant behaves correctly on iOS:
  - Xcode 16
  - Xcode 16 + UIDesignRequiresCompatibility
  - Xcode 26
  - Xcode 26 + UIDesignRequiresCompatibility
2. Test if app does not crash on Android when `isLiquidGlassAvailable`
is checked and that the value is `false`

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)