[core] Add standardised logging from native to JS across Android and iOS (#39159)
# Why

Currently the options to log to the JS side are limited and different on
both iOS and Android. Android has `errorManager` in `appContext`. iOS
has legady `EXLogManager`, which still works but isn't sufficient. A
standardised way of logging to JS from both platforms is needed.

# How

Based on the ErrorManager implementation on Android, I've added
`JSLoggerModule`. On Android it is compatible with `ErrorManagerModule`.
I have also deprecated ErrorManagerModule instead of extending it,
because the naming doesn't fit (we can also log info debug, etc. now).

The iOS implementation is basically the same as Android. Both are based
on the existing `Logger` interfaces for their platforms.

Now on both platforms we can log to JS console via
`appContext.jsLogger.[debug/warn/error/trace/info](message)`

# Test Plan

Tested in BareExpo on Android and iOS.