fix(metro-config): Collapse `unknown`/`<anonymous>` stack frames (#39407)
# Why

This is to prevent a `readFileSync` error when stacks with `unknown` in
the stack trace aren't collapsed.

This probably shows a bit of a weakness in the current approach Metro
has in processing stack traces, as `readFileSync` is blindly called on
any stack frame and the error is logged. Logging an error when
desymbolicating an error is undesirable.

# How

For now, the easiest solution is to ignore `unknown` in stack traces.

# Test Plan

- Manually tested

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)