[ios][sdk-54][precompile] added support for USE_FRAMEWORKS with Expo
Some Expo modules are including React Native Core header files to override base RN classes in their public headers. This causes the swift interop generation to fail, since the RN headers are not fully modular - some include c++ code which fails when generating the module interface for these modules.

This isn't a problem for the module itself - but for the modules or targets that are using them. XCode will try to generate a public and modular interface for the module using the Objective-c headers - but when these contains c++ the process fails.

Untill we get a truly modular set of headers for React Native using USE_FRAMEWORKS will fail with React Native precompiled binaries. Building from source or building without USE_FRAMEWORKS does not fail since header files are detected through search paths only - without any checks for public/modularity.

This commit fixes this by disabling frameworks (forcing static linking) of the targets that uses RN headers in their own public headers (Expo, ExpoModulesCore, ReactAppDependencyProvider, expo-dev-menu).

To fix this we are inserting a pre_install hook in the pod file that fixes the issues with react-core-prebuilt:

- Added pre_install script in expo/scripts
- Added call to pre_install in Podfiles:
  - Bare-Expo/Podfile
  - expo-template-bare-minimum/ios/Podfile
  - native-tests/ios
  - paper-tester/ios (not really necessary since this is only a problem on the new arch).

Tested in mulitple scenarios:

With new Expo app using SDK-54 and the changes from this commit:

✅ Build without precompiled, no USE_FRAMEWORKS
✅ Build without precompiled, USE_FRAMEWORKS=static
✅ Build without precompiled, USE_FRAMEWORKS=dynamic
✅ Build with precompiled, no USE_FRAMEWORKS
✅ Build with precompiled, USE_FRAMEWORKS=static
✅ Build with precompiled, USE_FRAMEWORKS=dynamic